Move BuddiesComboBox to utils
