+++ pynicotine/gtkgui/frame.py
@@ -72,6 +72,7 @@ from pynicotine.gtkgui.userinfo import UserInfo
 from pynicotine.gtkgui.userinfo import UserTabs
 from pynicotine.gtkgui.userlist import UserList
 from pynicotine.gtkgui.utils import AppendLine
+from pynicotine.gtkgui.utils import BuddiesComboBox
 from pynicotine.gtkgui.utils import Humanize
 from pynicotine.gtkgui.utils import HumanSpeed
 from pynicotine.gtkgui.utils import ImageLabel
@@ -88,51 +89,6 @@ from pynicotine.utils import version
 
 gi.require_version('Gtk', '3.0')
 gi.require_version('Gdk', '3.0')
-
-
-class BuddiesComboBox:
-
-    def __init__(self, frame, ComboBox):
-
-        self.frame = frame
-
-        self.items = {}
-
-        self.combobox = ComboBox
-
-        self.store = gtk.ListStore(gobject.TYPE_STRING)
-        self.combobox.set_model(self.store)
-        self.combobox.set_entry_text_column(0)
-
-        self.store.set_default_sort_func(lambda *args: -1)
-        self.store.set_sort_column_id(-1, gtk.SortType.ASCENDING)
-
-        self.combobox.show()
-
-    def Fill(self):
-
-        self.items.clear()
-        self.store.clear()
-
-        self.items[""] = self.store.append([""])
-
-        for user in self.frame.np.config.sections["server"]["userlist"]:
-            self.items[user[0]] = self.store.append([user[0]])
-
-        self.store.set_sort_column_id(0, gtk.SortType.ASCENDING)
-
-    def Append(self, item):
-
-        if item in self.items:
-            return
-
-        self.items[item] = self.combobox.get_model().append([item])
-
-    def Remove(self, item):
-
-        if item in self.items:
-            self.combobox.get_model().remove(self.items[item])
-            del self.items[item]
 
 
 class NicotineFrame:
